mark classes as final and methods and properties as private
We add interfaces for final classes so that they can be mocked in testing.
We chose to use the Interface suffix here to avoid name clashes, rather than renaming the existing PHP classes.
The interfaces are only relevant for unit testing, we do not expect people to provide their own implementations.

HttpClientPoolItem is purely internal and we therefore don't want an interface for it.
It is marked with @final and explained to not be meant to extend.